matching & generation via Thompson-Earley
