+++ package.yaml
@@ -27,6 +27,7 @@ dependencies:
 - distributive >= 0.6 && < 1
 - lens >= 5.0 && < 6
 - MemoTrie >= 0.6 && < 1
+- QuickCheck >= 2.14 && < 3
 - mtl >= 2.2 && < 3
 - profunctors >= 5.6 && < 6
 - tagged >= 0.8 && < 1
