+++ src/Control/Lens/Grammar/BackusNaur.hs
@@ -19,13 +19,10 @@ module Control.Lens.Grammar.BackusNaur
   , liftBnf0
   , liftBnf1
   , liftBnf2
-    -- * Matching
-  , Matching (..)
   , diffB
   ) where
 
 import Control.Lens
-import Control.Lens.Extras
 import Control.Lens.Grammar.Kleene
 import Control.Lens.Grammar.Token
 import Control.Lens.Grammar.Symbol
@@ -33,12 +30,6 @@ import Data.Bifunctor.Joker
 import Data.Coerce
 import Data.Foldable
 import Data.Function
-import qualified Data.IntMap.Strict as IntMap
-import Data.IntMap.Strict (IntMap)
-import qualified Data.IntSet as IntSet
-import Data.IntSet (IntSet)
-import qualified Data.Map.Strict as Map
-import Data.Map.Strict (Map)
 import Data.MemoTrie
 import qualified Data.Set as Set
 import Data.Set (Set)
@@ -113,11 +104,6 @@ liftBnf2
 liftBnf2 f (Bnf start0 rules0) (Bnf start1 rules1) =
   Bnf (f start0 start1) (Set.map coerce rules0 <> Set.map coerce rules1)
 
--- | Does a word match a pattern?
-class Matching word pattern | pattern -> word where
-  (=~) :: word -> pattern -> Bool
-  infix 2 =~
-
 {- |
 The [Brzozowski derivative]
 (https://dl.acm.org/doi/pdf/10.1145/321239.321249) of a
@@ -177,228 +163,6 @@ rulesNamed nameX = foldl' (flip inserter) Set.empty where
   inserter (nameY,y) =
     if nameX == nameY then Set.insert y else id
 
--- | A state in the Earley-extended Thompson transducer for a `Bnf`.
--- @EarleyTerminal cls ds@ matches on a token class and transitions to @ds@.
--- @EarleyNonTerminal name ds@ is a call point for rule @name@; after @name@
--- completes, control flows to @ds@. @EarleyEmit name@ is the final state
--- for rule @name@ and triggers completion during Earley closure.
-data EarleyState token
-  = EarleyTerminal (TokenClass token) IntSet
-  | EarleyNonTerminal String IntSet
-  | EarleyEmit String
-
-data EarleyTransducer token = EarleyTransducer
-  { earleyStates :: IntMap (EarleyState token)
-  , earleyRules :: Map String (IntSet, Bool)
-  , earleyAcceptId :: Int
-  , earleyStartStates :: IntSet
-  }
-
-compileEarley :: Bnf (RegEx token) -> EarleyTransducer token
-compileEarley (Bnf start rules) = EarleyTransducer
-  { earleyStates = IntMap.fromList allStates
-  , earleyRules = Map.fromList
-      [ ( n
-        , ( Map.findWithDefault IntSet.empty n firstsMap
-          , Map.findWithDefault False n nullMap
-          )
-        )
-      | n <- Map.keys ruleMap
-      ]
-  , earleyAcceptId = earleyAcceptId0
-  , earleyStartStates = startStates
-  }
-
-  where
-
-    ruleMap = foldr
-      (\(n, r) -> Map.insertWith (++) n [r]) Map.empty (toList rules)
-
-    rexNullable nm = \case
-      SeqEmpty -> True
-      NonTerminal n -> Map.findWithDefault False n nm
-      Sequence x y -> rexNullable nm x && rexNullable nm y
-      KleeneStar _ -> True
-      KleeneOpt _ -> True
-      KleenePlus x -> rexNullable nm x
-      RegExam (Alternate x y) -> rexNullable nm x || rexNullable nm y
-      RegExam (OneOf _) -> False
-      RegExam (NotOneOf _ _) -> False
-
-    iterNull nm =
-      let nm' = Map.mapWithKey
-            (\n _ -> any (rexNullable nm) (Map.findWithDefault [] n ruleMap)) nm
-      in if nm == nm' then nm else iterNull nm'
-
-    nullMap = iterNull (Map.map (const False) ruleMap)
-
-    ruleNames = Map.keys ruleMap
-
-    earleyAcceptId0 = 0
-
-    (finalMap, nextIdAfterFinals) =
-      foldl' alloc (Map.empty, earleyAcceptId0 + 1) ruleNames
-      where alloc (m, i) n = (Map.insert n i m, i + 1)
-
-    finalStatesList = [(finalMap Map.! n, EarleyEmit n) | n <- ruleNames]
-
-    (rulesStatesList, firstsMap, nextIdAfterRules) =
-      foldl' compileRule ([], Map.empty, nextIdAfterFinals) (Map.toList ruleMap)
-      where
-        compileRule (sts, fm, nid) (name, prods) =
-          let finalId = finalMap Map.! name
-              (newSts, newFirsts, nid') =
-                foldl' compileProd ([], IntSet.empty, nid) prods
-              compileProd (s, fs, i) prod =
-                let (f, st, i', _) =
-                      goEarley prod i (IntSet.singleton finalId)
-                in (s <> st, fs <> f, i')
-          in (sts <> newSts, Map.insert name newFirsts fm, nid')
-
-    (startFirsts, startStatesRaw, _, startBypass) =
-      goEarley start nextIdAfterRules (IntSet.singleton earleyAcceptId0)
-
-    startStates =
-      startFirsts <> bypassStates startBypass (IntSet.singleton earleyAcceptId0)
-
-    allStates = finalStatesList <> rulesStatesList <> startStatesRaw
-
-    bypassStates True = id
-    bypassStates False = const IntSet.empty
-
-    goEarley rex nextId dests = case rex of
-        SeqEmpty -> (IntSet.empty, [], nextId, True)
-        NonTerminal name ->
-          ( IntSet.singleton nextId
-          , [(nextId, EarleyNonTerminal name dests)]
-          , nextId + 1
-          , Map.findWithDefault False name nullMap
-          )
-        Sequence rex0 rex1 ->
-          let
-            (firsts1, states1, nextId1, bypass1) = goEarley rex1 nextId dests
-            (firsts0, states0, nextId0, bypass0) =
-              goEarley rex0 nextId1 (firsts1 <> bypassStates bypass1 dests)
-          in
-            ( firsts0 <> bypassStates bypass0 firsts1
-            , states0 <> states1
-            , nextId0
-            , bypass0 && bypass1
-            )
-        KleeneStar rex0 ->
-          let
-            (firsts, states, nextId', _) = goEarley rex0 nextId (firsts <> dests)
-          in
-            (firsts, states, nextId', True)
-        KleeneOpt rex0 ->
-          let
-            (firsts, states, nextId', _) = goEarley rex0 nextId dests
-          in
-            (firsts, states, nextId', True)
-        KleenePlus rex0 ->
-          let
-            (firsts, states, nextId', bypass) = goEarley rex0 nextId (firsts <> dests)
-          in
-            (firsts, states, nextId', bypass)
-        RegExam (OneOf chars)
-          | Set.null chars -> (IntSet.empty, [], nextId, False)
-          | otherwise ->
-              ( IntSet.singleton nextId
-              , [(nextId, EarleyTerminal (TokenClass (OneOf chars)) dests)]
-              , nextId + 1
-              , False
-              )
-        RegExam (NotOneOf chars catTest) ->
-          ( IntSet.singleton nextId
-          , [(nextId, EarleyTerminal (TokenClass (NotOneOf chars catTest)) dests)]
-          , nextId + 1
-          , False
-          )
-        RegExam (Alternate rex0 rex1) ->
-          let
-            (firsts1, states1, nextId1, bypass1) = goEarley rex1 nextId dests
-            (firsts0, states0, nextId0, bypass0) = goEarley rex0 nextId1 dests
-          in
-            ( firsts0 <> firsts1
-            , states0 <> states1
-            , nextId0
-            , bypass0 || bypass1
-            )
-
-matchEarley :: Categorized token => [token] -> EarleyTransducer token -> Bool
-matchEarley word et = IntSet.member 0 acceptOrigins
-  where
-
-    initialE0 = IntMap.fromList
-      [ (s, IntSet.singleton 0) | s <- IntSet.toList (earleyStartStates et) ]
-
-    sets0 = IntMap.singleton 0 initialE0
-
-    sets0closed = closureAt 0 sets0
-
-    (finalSets, n) = runInput 0 sets0closed word
-
-    runInput j ss [] = (ss, j)
-    runInput j ss (x : xs) =
-      let scanned = scanFrom j x ss
-          ss' = IntMap.insert (j + 1) scanned ss
-          closed = closureAt (j + 1) ss'
-      in runInput (j + 1) closed xs
-
-    en = IntMap.findWithDefault IntMap.empty n finalSets
-
-    acceptOrigins = IntMap.findWithDefault IntSet.empty (earleyAcceptId et) en
-
-    scanFrom j input ss = IntMap.foldrWithKey advance IntMap.empty e_j
-      where
-        e_j = IntMap.findWithDefault IntMap.empty j ss
-        advance s origs acc = case IntMap.lookup s (earleyStates et) of
-          Just (EarleyTerminal cls ds) | tokenClass cls input ->
-            IntSet.foldr
-              (\d -> IntMap.insertWith IntSet.union d origs) acc ds
-          _ -> acc
-
-    closureAt j initialSets = loop initialWork initialSets
-      where
-        initialE = IntMap.findWithDefault IntMap.empty j initialSets
-        initialWork =
-          [ (s, i) | (s, os) <- IntMap.toList initialE, i <- IntSet.toList os ]
-        loop [] ss = ss
-        loop ((s, i) : rest) ss = case IntMap.lookup s (earleyStates et) of
-          Just (EarleyNonTerminal name ds) ->
-            let (firsts, isNull) = Map.findWithDefault
-                  (IntSet.empty, False) name (earleyRules et)
-                predItems = [(f, j) | f <- IntSet.toList firsts]
-                nullItems =
-                  if isNull then [(d, i) | d <- IntSet.toList ds] else []
-                (ss', new) = addEarleyItems j (predItems <> nullItems) ss
-            in loop (new <> rest) ss'
-          Just (EarleyEmit name) ->
-            let e_i = IntMap.findWithDefault IntMap.empty i ss
-                completions =
-                  [ (d, i')
-                  | (t, os) <- IntMap.toList e_i
-                  , Just (EarleyNonTerminal n' ds) <- [IntMap.lookup t (earleyStates et)]
-                  , n' == name
-                  , i' <- IntSet.toList os
-                  , d <- IntSet.toList ds
-                  ]
-                (ss', new) = addEarleyItems j completions ss
-            in loop (new <> rest) ss'
-          _ -> loop rest ss
-
-    addEarleyItems j items ss = foldl' ins (ss, []) items
-      where
-        ins (acc, new) (state, origin) =
-          let e_j = IntMap.findWithDefault IntMap.empty j acc
-              os = IntMap.findWithDefault IntSet.empty state e_j
-          in if IntSet.member origin os
-            then (acc, new)
-            else
-              let e_j' = IntMap.insert state (IntSet.insert origin os) e_j
-                  acc' = IntMap.insert j e_j' acc
-              in (acc', (state, origin) : new)
-
 -- instances
 instance (Ord rule, NonTerminalSymbol rule)
   => BackusNaurForm (Bnf rule) where
@@ -438,11 +202,3 @@ instance (Ord rule, Monoid rule) => Monoid (Bnf rule) where
   mempty = liftBnf0 mempty
 instance (Ord rule, Semigroup rule) => Semigroup (Bnf rule) where
   (<>) = liftBnf2 (<>)
-instance Categorized token
-  => Matching [token] (Bnf (RegEx token)) where
-    word =~ bnf = matchEarley word (compileEarley bnf)
-instance Categorized token
-  => Matching [token] (RegEx token) where
-    word =~ pattern = word =~ liftBnf0 pattern
-instance Matching s (APrism s t a b) where
-  word =~ pattern = is pattern word
