+++ src/Data/Profunctor/Grammar.hs
@@ -27,6 +27,7 @@ import Control.Lens
 import Control.Lens.Extras
 import Control.Lens.Grammar.BackusNaur
 import Control.Lens.Grammar.Kleene
+import Control.Lens.Grammar.Matching
 import Control.Lens.Grammar.Symbol
 import Control.Lens.Grammar.Token
 import Control.Monad
